memory allocation logic more safety

+++ lib/SPVM/Builder/include/spvm_runtime_api.h
@@ -123,8 +123,8 @@ void SPVM_RUNTIME_API_leave_scope(SPVM_ENV* env, int32_t original_mortal_stack_t
 int32_t SPVM_RUNTIME_API_call_sub(SPVM_ENV* env, int32_t sub_id, SPVM_VALUE* stack);
 int32_t SPVM_RUNTIME_API_call_sub_vm(SPVM_ENV* env, int32_t sub_id, SPVM_VALUE* stack);
 
-void* SPVM_RUNTIME_API_alloc_memory_block_zero(SPVM_ENV* env, size_t byte_size);
+void* SPVM_RUNTIME_API_alloc_memory_block_zero(SPVM_ENV* env, int64_t byte_size);
 void SPVM_RUNTIME_API_free_memory_block(SPVM_ENV* env, void* block);
-void* SPVM_RUNTIME_API_safe_malloc_zero(size_t byte_size);
+void* SPVM_RUNTIME_API_safe_malloc_zero(int64_t byte_size);
 
 #endif
